pyproject.toml
[tool.poetry]
name = "elixir-ai"
version = "0.1.0"
description = ""
authors = ["Andrew Chang <andrew@acalabs.xyz>"]
readme = "README.md"
packages = [
    { include = "elixir" },
]

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.9.5"
opentelemetry-sdk = "^1.24.0"
python-dotenv = "^1.0.1"
traceloop-sdk = "^0.21.0"
aioboto3 = "^13.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
